repository: use the older version tag of a commit and use previous non-empty commit
If a commit has both v0.52.0 and v0.52.1 tag, pick
v0.52.0 tag to compute version because a partial v0.52.1
build would be empty and then not produce a dpk,
meaning doing a partial build with v0.52.1 reference
would rely on non existant v0.52.1 dpk instead of
the v0.52.0 one.

If a commit produces an empty package, use the previous
commit not producing an empty package, if the reference
name is a version tag, use the previous version tag not
producing an empty package.