parallelism: Attempt to get number of cpu from `nproc --all` or fallback on psutil.cpu_count()
On Linux, psutil.cpu_count() and `nproc` may only return the
amount of currently enabled core while the system may disable
or disable them according to the current load, meaning the
reported cpu count may be far lower to what's available.
If we don't ask for the amount of available the present software
may limit itself to a small amount of cores if the system was
idling at the time the present software is called.
The `nproc --all` command is known to report the amount of
available cores.