urcheon: new layout with pkg, build/pakdir/pkg, build/pkg
Instead of using those folders:

  src/name_src.dpkdir
  build/test/name_test.dpkdir
  build/pkg/name_<version>.dpk

We now use those folders:

  pkg/name_src.dpkdir
  build/pakdir/pkg/name_test.dpkdir
  build/pkg/name_<version>.dpk

It means the base folder is always `pkg`.

It makes compatibility with legacy idTech tools is better
as they expect the same folder name everywhere.

It makes also easier to implement Urcheon compatibility
in third party software.

This also make possible to implement compatibility
with arbitrary base folder name (like `data` for Xonotic),
or even with Quake3-like mods in the future with multiple base
folders. This is not a priority but the fact that design
makes it possible means it's a better design.

Right now this is good enough for Unvanquished and
Dæmon based games.

Data repositories like UnvanquishedAssets or InterstellarOasis
have to be migrated this way:

  [ -d pkg ] && mv pkg pkg.old
  git mv src pkg
  mkdir build/pakdir
  mv build/test build/pakdir/pkg

Urcheon doesn't provide any compatibility layer with the
previous layout, as the cost of that complexity is considered
too high compared to the amounts of repositories to migrates.

A typical build is done this way:

  urcheon prepare pkg/*.dpkdir
  urcheon build pkg/*.dpkdir
  urcheon package pkg/*.dpkdir

The engine can then be run with `-pakpath build/pakdir/pkg` to load dpkdir
or be run with `-pakpath install/pkg` to load dpk.

The `--pakprefix` option is now working like the
engine `fs_pakprefixes` cvar does, see:

- https://github.com/DaemonEngine/Daemon/pull/964

Some Urcheon-specific environment variables are now prefixed with `URCHEON_`.